Move clear clipboard logic to its own file
